Implement event management handlers: create, update, delete, and get events
